Texture viewer now usable.
Allows specifying the file to view on the command line so can be used with windows open/double-click. Supports viewing of images with transparency (32 bit pngs and tgas for example). Info dialog no longer on top of where the image is displayed. Files are sorted alphabetically. Font always loads by using the program dir rather than the working dir.